Move legacy modules to sleap_mot/_legacy/
- Move tracker.py, track_instance.py, feature_tracker.py, candidates/ to _legacy/
- Update internal imports within legacy modules
- Update old tests to import from _legacy
- Clean sleap_mot/__init__.py to expose only current API
