+++ sleap_mot/__init__.py
@@ -1,12 +1,5 @@
-"""This module exposes all high level APIs for sleap-io."""
+"""sleap-mot: Standalone multi-object tracking for the SLEAP ecosystem."""
 
 from sleap_mot.version import __version__
-from sleap_mot import tracker, track_instance, candidates, utils
-from sleap_mot.tracker import Tracker, FlowShiftTracker
-from sleap_mot.track_instance import (
-    TrackedInstanceFeature,
-    TrackInstanceLocalQueue,
-    TrackInstances,
-)
-from sleap_mot.candidates.local_queues import LocalQueueCandidates
-from sleap_mot.candidates.fixed_window import FixedWindowCandidates
+from sleap_mot import utils, io
+from sleap_mot.io import SLPTFile
