+++ sleap_mot/_legacy/__init__.py
@@ -1,0 +1,14 @@
+"""Legacy tracking modules (deprecated).
+
+These modules implement the original sleap-mot tracking API and are no longer
+actively developed. They are preserved here for reference and backward
+compatibility with older scripts.
+
+The current tracking system lives in sleap_mot.tracking.
+
+Modules:
+    tracker: Original Tracker and FlowShiftTracker classes
+    track_instance: TrackedInstanceFeature, TrackInstances, TrackInstanceLocalQueue
+    feature_tracker: Original RFIDFeatureTracker (superseded by tracking.feature_tracking)
+    candidates: FixedWindowCandidates, LocalQueueCandidates
+"""
