+++ sleap_mot/_legacy/candidates/fixed_window.py
@@ -1,7 +1,7 @@
 """Module to generate Fixed window candidates."""
 
 from typing import Optional, List, Deque, Union
-from sleap_mot.track_instance import TrackInstances, TrackedInstanceFeature
+from sleap_mot._legacy.track_instance import TrackInstances, TrackedInstanceFeature
 import sleap_io as sio
 from collections import deque
 import numpy as np
